pyproject.toml
[project]
name = "trading-bot"
version = "0.1.0"
description = ""
authors = [
    {name = "Paulo Roberto, paulor1610@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "pandas (>=2.2.3,<3.0.0)",
    "ccxt (>=4.4.53,<5.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "plotext (>=5.3.2,<6.0.0)",
    "ta (>=0.11.0,<0.12.0)",
    "colorama (>=0.4.6,<0.5.0)"

]
package-mode = false
[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"